Refactor code: Parser, Storage, TasksList, and Ui should be in Duke's utility package
